:hammer: Move EndpointInfoFactory to commonMain for multiplatform reuse
