--- composeApp/src/desktopMain/kotlin/com/crosspaste/app/DesktopEndpointInfoFactory.kt
@@ -1,31 +1,0 @@
-package com.crosspaste.app
-
-import com.crosspaste.dto.sync.EndpointInfo
-import com.crosspaste.net.PasteServer
-import com.crosspaste.platform.getPlatform
-import com.crosspaste.realm.sync.HostInfo
-import com.crosspaste.utils.DeviceUtils
-import com.crosspaste.utils.getNetUtils
-
-class DesktopEndpointInfoFactory(
-    private val deviceUtils: DeviceUtils,
-    private val pasteServer: Lazy<PasteServer<*, *>>,
-) : EndpointInfoFactory {
-
-    private val netUtils = getNetUtils()
-
-    override fun createEndpointInfo(hostInfoFilter: (HostInfo) -> Boolean): EndpointInfo {
-        val platform = getPlatform()
-        val port = pasteServer.value.port()
-        val hostInfoList = netUtils.getHostInfoList(hostInfoFilter)
-        val deviceName = deviceUtils.getDeviceName()
-        val deviceId = deviceUtils.getDeviceId()
-        return EndpointInfo(
-            deviceId = deviceId,
-            deviceName = deviceName,
-            platform = platform,
-            hostInfoList = hostInfoList,
-            port = port,
-        )
-    }
-}
